14/02/2021 Upstream merge

+++ src/lib/l1/ECL_L1_Pos_Controller.hpp
@@ -143,8 +143,8 @@ public:
 	 *
 	 * @return sets _lateral_accel setpoint
 	 */
-	void navigate_waypoints(const matrix::Vector2f &vector_A, const matrix::Vector2f &vector_B,
-				const matrix::Vector2f &vector_curr_position, const matrix::Vector2f &ground_speed);
+	void navigate_waypoints(const matrix::Vector2d &vector_A, const matrix::Vector2d &vector_B,
+				const matrix::Vector2d &vector_curr_position, const matrix::Vector2f &ground_speed);
 
 	/**
 	 * Navigate on an orbit around a loiter waypoint.
@@ -154,7 +154,7 @@ public:
 	 *
 	 * @return sets _lateral_accel setpoint
 	 */
-	void navigate_loiter(const matrix::Vector2f &vector_A, const matrix::Vector2f &vector_curr_position, float radius,
+	void navigate_loiter(const matrix::Vector2d &vector_A, const matrix::Vector2d &vector_curr_position, float radius,
 			     int8_t loiter_direction, const matrix::Vector2f &ground_speed_vector);
 
 	/**
@@ -235,7 +235,7 @@ private:
 	 * @param wp The point to convert to into the local coordinates, in WGS84 coordinates
 	 * @return The vector in meters pointing from the reference position to the coordinates
 	 */
-	matrix::Vector2f get_local_planar_vector(const matrix::Vector2f &origin, const matrix::Vector2f &target) const;
+	matrix::Vector2f get_local_planar_vector(const matrix::Vector2d &origin, const matrix::Vector2d &target) const;
 
 	/**
 	 * Update roll angle setpoint. This will also apply slew rate limits if set.
